+++ tests/evaluation/test_metrics.py
@@ -1,10 +1,10 @@
 
 import pytest
 import numpy as np
-from dtaianomaly.evaluation import Precision, Recall, FBeta, AreaUnderPR, AreaUnderROC, ThresholdMetric
+from dtaianomaly.evaluation import *
 from dtaianomaly.thresholding import FixedCutoff
 
-binary_metrics = [Precision(), Recall(), FBeta()]
+binary_metrics = [Precision(), Recall(), FBeta(), PointAdjustedPrecision(), PointAdjustedRecall(), PointAdjustedFBeta()]
 proba_metrics = [AreaUnderROC(), AreaUnderPR(), ThresholdMetric(FixedCutoff(0.5), Precision())]